-

NOMADjl.init — Method.

init(".../nomad3.9.1")

load NOMAD libraries and create C++ class and function needed to handle NOMAD optimization process.

This function has to be called once before using runopt. Moreover, it should not be called more than once (the C++ class and function cannot be overwritten).

The only argument is a String containing the path to NOMAD3.9.1 folder.

NOMADjl.parameters — Type.

parameters

mutable struct containing the options of the optimization process.

Best is to construct one with p=parameters() and then modify its attributes one after another. For deeper information concerning these attributes, please refer to the NOMAD documentation. the names of these options are the same in NOMADjl as in NOMAD.

#Attributes :

- dimension::Int64 : Dimension of the problem and
size of the argument given to eval.
2 by default.

- output_types::Vector{String} : A vector containing
String objects that define the types of outputs returned
by eval (the order is important) :
    > "OBJ" : objective value to be minimized
    > "EB" : extreme barrier constraint
    > "PB" : progressive barrier constraint
    > "F" : filter approach constraint
    > "PEB" : hybrid constraint EB/PB
    > ... (see NOMAD documentation)
["OBJ","EB"] by default.

- display_all_eval::Bool : if false, only evaluations
that allow to improve the current state are displayed.
false by default.

- display_stats::String : String defining the way outputs
are displayed (it should not contain any quotes). Here are
examples of keywords that can be used :
    > bbe : black box evaluations
    > obj : objective function value
    > sol : solution
    > bbo : black box outputs
    > mesh_index : mesh mesh_index
    > mesh_size : mesh size parameter
    > ... (see NOMAD documentation)
"bbe ( sol ) obj" by default.

- x0::Vector{Float} : Initialization point for
NOMAD. Its size needs to be equal to dimension.
[0,0] by default.

- lower_bound::Vector{Number} : Lower bound for
each coordinate of the state.
[-100,-100] by default.

- upper_bound::Vector{Number} : Upper bound for
each coordinate of the state.
[100, 100] by default.

- max_bb_eval::Int : Maximum of calls to eval
allowed.
100 by default.

- display_degree::Int : Integer between 0 and 3
that sets the level of display.
2 by default.

- solution_file::String : Name of the generated
output file containing the returned minimum.
"sol.txt" by default.

NOMADjl.runopt — Method.

runopt(eval,param)

Run NOMAD with settings defined by param and an optimization problem defined by eval.

init has to be called before using runopt.

#Arguments :

- eval::Function : a function of the form :

	(count_eval,bb_outputs)=eval(x)

	bb_outputs being a vector{Float64} containing
	the values of objectives functions and constraints
	for a given input vector x. NOMAD will seak for
	minimizing the objective functions and keeping
	constraints inferior to 0. count_eval is a
	boolean equal to true if the evaluation has
	to be taken into account by NOMAD.

- param::Parameters : an instance of the julia mutable
	struct Parameters of which the attributes are the
	settings of the optimization process (dimension,
	output types, display options, bounds, etc.).

This function first wraps eval with a julia function eval wrap that takes a C-double[] as argument and returns a C-double[]. Then it converts all param attributes into C++ variables and calls the C++ function cpp main previously defined by init_NOMADjl.