Results

Results

Main results from a NOMAD optimization process are stored in an object of the type described below.

NOMAD.results — Type.

results

mutable struct containing info about a NOMAD run, returned by the method runopt(eval,param).

To display the info contained in a object result, use :

disp(result)

Attributes :

best_feasible::Vector{Float64} :

Feasible point found by NOMAD that best minimizes the objective function.

bbo_best_feasible::Vector{Float64} :

Outputs of eval(x) for the best feasible point.

best_infeasible::Vector{Float64} :

Infeasible point found by NOMAD that best minimizes the objective function.

bbo_best_infeasible::Vector{Float64} :

outputs of eval(x) for the best infeasible point.

bb_eval::Int64 :

Number of eval(x) evaluations

source