Run the optimization

A NOMAD optimization process can be run by using the nomad() method described below.

NOMAD.nomad — Method

nomad(eval::Function,param::nomadParameters)

-> Run NOMAD with settings defined by param and an optimization problem defined by eval.

-> Display stats from NOMAD in the REPL.

-> return an object of type nomadResults that contains info about the run.

Arguments :

eval::Function

a function of the form :

(success,counteval,bboutputs)=eval(x::Vector{Number})

bb_outputs being a vector{Number} containing the values of objective function and constraints for a given input vector x. NOMAD will seak to minimize the objective function and keeping constraints inferior to 0.

success is a Bool set to false if the evaluation failed.

count_eval is a Bool equal to true if the black box evaluation counting has to be incremented. Note that statistic sums and averages are updated only if count_eval is equal to true.

param::nomadParameters

An object of type nomadParameters of which the attributes are the settings of the optimization process (dimension, output types, display options, bounds, etc.).

Example :

using NOMAD

function eval(x) f=x[1]^2+x[2]^2 c=1-x[1] success=true counteval=true bboutputs = [f,c] return (success,counteval,bboutputs) end

param = nomadParameters([5,5],["OBJ","EB"]) #=first element of bb_outputs is the objective function ("OBJ"), second is a constraint treated with the Extreme Barrier method ("EB"). Initial point for optimization is [5,5]=#

result = nomad(eval,param)

disp(result)