Run Optimization

Run the optimization

A NOMAD optimization process can be run by using the runopt method described below.

NOMAD.runopt — Method.

runopt(eval::Function,param::parameters)

-> Run NOMAD with settings defined by param and an optimization problem defined by eval(x).

-> Display stats from NOMAD in the REPL.

-> return an object of type results that contains info about the run.

Arguments :

eval::Function

a function of the form :

(count_eval,bb_outputs)=eval(x::Vector{Float64})

bb_outputs being a vector{Float64} containing the values of objective function and constraints for a given input vector x. NOMAD will seak for minimizing the objective function and keeping constraints inferior to 0. count_eval is a Bool equal to true if the evaluation has to be taken into account by NOMAD.

param::parameters

An object of type Parameters of which the attributes are the settings of the optimization process (dimension, output types, display options, bounds, etc.).

Example :

using NOMAD

function eval(x)
    f=x[1]^2+x[2]^2
    c=1-x[1]
    count_eval=true
	bb_outputs = [f,c]
    return (count_eval,bb_outputs)
end

param = parameters()
param.dimension = 2
param.output_types = ["OBJ","EB"] #=first element of bb_outputs is the
	objective function, second is a constraint treated with the Extreme
	Barrier method=#
param.x0 = [3,3] #Initial state for the optimization process

result = runopt(eval,param)

source