Parameters setting

The settings of a NOMAD optimization process must be entered in an object of the type described below.

NOMAD.NomadProblem — Type

NomadProblem(nb_inputs::Int, nb_outputs::Int, output_types::Vector{String}, eval_bb::Function;
             input_types::Vector{String} = ["R" for i in 1:nb_inputs],
             granularity::Vector{Float64} = zeros(Float64, nb_inputs),
             lower_bound::Vector{Float64} = -Inf * ones(Float64, nb_inputs),
             upper_bound::Vector{Float64} = Inf * ones(Float64, nb_inputs),
             A::Union{Nothing, Matrix{Float64}} = nothing,
             b::Union{Nothing, Vector{Float64}} = nothing,
             min_mesh_size::Vector{Float64} = zeros(Float64, nb_inputs),
             b::Union{Nothing, Vector{Float64}} = nothing,
             initial_mesh_size::Vector{Float64} = Float64[])

Struct containing the main information needed to solve a blackbox problem by the Nomad Software.

Attributes:

nb_inputs::Int:

Number of inputs of the blackbox. Is required to be > 0.

No default, needs to be set.

nb_outputs::Int:

Number of outputs of the blackbox. Is required to be > 0.

No default, needs to be set.

output_types::Vector{String}:

A vector containing String objects that define the types of outputs returned by eval_bb (the order is important) :

String	Output type

"OBJ"	objective value to be minimized

"PB"	progressive barrier constraint

"EB"	extreme barrier constraint

No default value, needs to be set.

eval_bb::Function:

A function of the form :

    (success, count_eval, bb_outputs) = eval(x::Vector{Float64})

bb_outputs being a Vector{Float64} containing the values of objective function and constraints for a given input vector x. NOMAD will seek to minimize the objective function and keeping constraints inferior to 0.

success is a Bool set to false if the evaluation failed.

count_eval is a Bool equal to true if the blackbox evaluation counting has to be incremented.

input_types::Vector{String}:

A vector containing String objects that define the types of inputs to be given to eval_bb (the order is important) :

String	Input type

"R"	Real/Continuous

"B"	Binary

"I"	Integer

all R by default.

granularity::Vector{Float64}:

The granularity of input variables, that is to say the minimum variation authorized for these variables. A granularity of 0 corresponds to a real variable.

By default, 0 for real variables, 1 for integer and binary ones.

min_mesh_size::Vector{Float64}:

The minimum mesh size to reach allowed by each input variable. When a variable decreases below the threshold, the algorithm stops.

By default, 0 (which corresponds to the Nomad software tolerance).

initial_mesh_size::Vector{Float64}:

The initial mesh size set for each input variable. Can be adjusted if the granularity is set.

Empty by default.

lower_bound::Vector{Float64}:

Lower bound for each coordinate of the blackbox input. -Inf * ones(Float64, nb_inputs), by default.

upper_bound::Vector{Float64}:

Upper bound for each coordinate of the blackbox input.

Inf * ones(Float64, nb_inputs), by default.

A::Union{Nothing, Matrix{Float64}}:

Matrix A in the potential equality constraints Ax = b, where x are the inputs of the blackbox. A must have more columns than lines. If defined, the granularity parameters should be set to default value, i.e. 0.

nothing, by default.

b::Union{Nothing, Vector{Float64}}:

Vector b in the potential equality constraints Ax=b, where x are the inputs of the blackbox. b must be defined when A is defined. In this case, dimensions must match.

nothing, by default.

options::NomadOptions

Nomad options that can be set before running the optimization process.

-> max_cache_size::Int:

Maximum number of points stored in the cache.

Inf by default.

-> display_degree::Int:

Integer between 0 and 3 that sets the level of display.

-> display_all_eval::Bool:

If false, only evaluations that allow to improve the current state are displayed.

false by default.

-> display_infeasible::Bool:

If true, display best infeasible values reached by Nomad until the current step.

false by default.

-> display_unsuccessful::Bool:

If true, display evaluations that are unsuccessful.

false by default.

-> display_stats::Bool:

A vector containing String objects that define the statistics to display when the algorithm is running.

String	Display Statistics Arguments

"BBE"	Blackbox evaluations

"BBO"	Blackbox outputs

"CACHE_HITS"	Cache hits

"CACHE_SIZE"	Cache size

"CONS_H"	Infeasibility (h) value

"EVAL"	Evaluations (includes cache hits)

"FEAS_BBE"	Feasible blackbox evaluations

"GEN_STEP"	Name of the step that generated

this point to evaluate

"H_MAX"	Max infeasibility (h) acceptable

"INF_BBE"	Infeasible blackbox evaluations

"ITER_NUM"	Iteration number in which this

evaluation was done

"MESH_INDEX"	Mesh index

"OBJ"	Objective function value

"PHASE_ONE_SUCC"	Success evaluations during phase one

phase

"SGTE"	Number of surrogate evaluations since

last reset

"SOL"	Current feasible iterate (displayed

in ())

"SUCCESS_TYPE"	Success type for this evaluation,

compared with the frame center

"TIME"	Real time in seconds

"TOTAL_SGTE"	Total number of surrogate evaluations

Empty by default.

-> max_bb_eval::Int:

Maximum of calls to eval_bb allowed. Must be positive.

20000 by default.

-> max_sgte_eval::Int:

Maximum of calls to surrogate models for each optimization of surrogate problem allowed. Must be positive.

1000 by default.

-> opportunistic_eval::Bool

If true, the algorithm performs an opportunistic strategy at each iteration.

true by default.

-> use_cache::Bool:

If true, the algorithm only evaluates one time a given input. Avoids to recalculate a blackbox value if this last one has already be computed.

true by default.

-> random_eval_sort::Bool:

If true, trial points are randomly shuffled before being evaluated.

false by default.

-> lh_search::Tuple{Int, Int}:

LH search parameters.

lh_search[1] is the lh_search_init parameter, i.e. the number of initial search points performed with Latin-Hypercube method.

0 by default.

lh_search[2] is the lh_search_iter parameter, i.e. the number of search points performed at each iteration with Latin-Hypercube method.

0 by default.

-> quad_model_search::Bool:

If true, the algorithm executes a quadratic model search strategy at each iteration. Deactivated when the number of variables is greater than 50.

true by default.

-> sgtelib_search::Bool:

If true, the algorithm executes a model search strategy at each iteration. Deactivated when the number of variables is greater than 50.

false by default.

-> sgtelib_model_trials::Int:

Maximum number of model search steps by iteration to try, before going to the poll step. Must be positive.

1 by default.

-> speculative_search::Bool:

If true, the algorithm executes a speculative search strategy at each iteration.

true by default.

-> speculative_search_max::Int:

Number of points to generate using the Mads speculative search (when opportunistic strategy). Must be positive.

1 by default.

-> nm_search::Bool:

If true, the algorithm executes a speculative search strategy at each iteration.

true by default.

-> nm_search_stop_on_success::Bool:

If true, the nm_search strategy stops opportunistically (as soon as a better point is found).

false by default.

-> max_time::Union{Nothing, Int}:

If defined, maximum clock time (in seconds) execution of the algorithm.

false by default.

-> linear_converter::String:

The type of converter to deal with linear equality constraints. Can be SVD or QR.

SVD by default.

-> linear_constraints_atol::Float64:

The tolerance accuracy that x0 must satisfy, when there are linear equality constraints, i.e. A * x0 = b.

0 by default.