Parameters setting

The settings of a NOMAD optimization process must be entered in an object of the type described below.

NOMAD.NomadProblem — Type

NomadProblem(nb_inputs::Int, nb_outputs::Int, output_types::Vector{String}, eval_bb::Function;
             input_types::Vector{String} = ["R" for i in 1:nb_inputs],
             granularity::Vector{Float64} = zeros(Float64, nb_inputs),
             lower_bound::Vector{Float64} = -Inf * ones(Float64, nb_inputs),
             upper_bound::Vector{Float64} = Inf * ones(Float64, nb_inputs),
             A::Union{Nothing, Matrix{Float64}} = nothing,
             b::Union{Nothing, Vector{Float64}} = nothing)

Struct containing the main information needed to solve a blackbox problem by the Nomad Software.

Attributes:

nb_inputs::Int:

Number of inputs of the blackbox. Is required to be > 0.

No default, needs to be set.

nb_outputs::Int:

Number of outputs of the blackbox. Is required to be > 0.

No default, needs to be set.

output_types::Vector{String}:

A vector containing String objects that define the types of outputs returned by eval_bb (the order is important) :

String	Output type

"OBJ"	objective value to be minimized

"PB"	progressive barrier constraint

"EB"	extreme barrier constraint

No default value, needs to be set.

eval_bb::Function:

A function of the form :

    (success, count_eval, bb_outputs) = eval(x::Vector{Float64})

bb_outputs being a Vector{Float64} containing the values of objective function and constraints for a given input vector x. NOMAD will seek to minimize the objective function and keeping constraints inferior to 0.

success is a Bool set to false if the evaluation failed.

count_eval is a Bool equal to true if the blackbox evaluation counting has to be incremented.

input_types::Vector{String}:

A vector containing String objects that define the types of inputs to be given to eval_bb (the order is important) :

String	Input type

"R"	Real/Continuous

"B"	Binary

"I"	Integer

all R by default.

granularity::Vector{Float64}:

The granularity of input variables, that is to say the minimum variation authorized for these variables. A granularity of 0 corresponds to a real variable.

By default, 0 for real variables, 1 for integer and binary ones.

lower_bound::Vector{Float64}:

Lower bound for each coordinate of the blackbox input. -Inf * ones(Float64, nb_inputs), by default.

upper_bound::Vector{Float64}:

Upper bound for each coordinate of the blackbox input.

Inf * ones(Float64, nb_inputs), by default.

A::Union{Nothing, Matrix{Float64}}:

Matrix A in the potential equality constraints Ax = b, where x are the inputs of the blackbox. A must have more columns than lines. If defined, the granularity parameters should be set to default value, i.e. 0.

nothing, by default.

b::Union{Nothing, Vector{Float64}}:

Vector b in the potential equality constraints Ax=b, where x are the inputs of the blackbox. b must be defined when A is defined. In this case, dimensions must match.

nothing, by default.

options::NomadOptions

Nomad options that can be set before running the optimization process.

-> display_degree::Int:

Integer between 0 and 3 that sets the level of display.

-> display_all_eval::Bool:

If false, only evaluations that allow to improve the current state are displayed.

false by default.

-> display_infeasible::Bool:

If true, display best infeasible values reached by Nomad until the current step.

false by default.

-> display_unsuccessful::Bool:

If true, display evaluations that are unsuccessful.

false by default.

-> max_bb_eval::Int:

Maximum of calls to eval_bb allowed. Must be positive.

20000 by default.

-> opportunistic_eval::Bool

If true, the algorithm performs an opportunistic strategy at each iteration.

true by default.

-> use_cache::Bool:

If true, the algorithm only evaluates one time a given input. Avoids to recalculate a blackbox value if this last one has already be computed.

true by default.

-> lh_search::Tuple{Int, Int}:

LH search parameters.

lh_search[1] is the lh_search_init parameter, i.e. the number of initial search points performed with Latin-Hypercube method.

0 by default.

lh_search[2] is the lh_search_iter parameter, i.e. the number of search points performed at each iteration with Latin-Hypercube method.

0 by default.

-> speculative_search::Bool:

If true, the algorithm executes a speculative search strategy at each iteration.

true by default.

-> nm_search::Bool:

If true, the algorithm executes a speculative search strategy at each iteration.

true by default.

-> linear_converter::String:

The type of converter to deal with linear equality constraints. Can be SVD or QR.

SVD by default.

-> linear_constraints_atol::Float64:

The tolerance accuracy that x0 must satisfy, when there are linear equality constraints, i.e. A * x0 = b.

0 by default.