Results

Results

Main results from a NOMAD optimization process are stored in an object of the type described below.

NOMAD.nomadResults — Type.

nomadResults

mutable struct containing info about a NOMAD run, returned by the method runopt(eval,param).

To display the info contained in a object result, use :

disp(result)

Attributes :

best_feasible::Vector{Float64} :

Feasible point found by NOMAD that best minimizes the objective function.

bbo_best_feasible::Vector{Float64} :

Outputs of eval(x) for the best feasible point.

best_infeasible::Vector{Float64} :

Infeasible point found by NOMAD that best minimizes the objective function.

bbo_best_infeasible::Vector{Float64} :

outputs of eval(x) for the best infeasible point.

bb_eval::Int64 :

Number of eval(x) evaluations

inter_states::Matrix{Float64} :

List of intermediate states evaluated during the optimization process. Lines correspond to successive states that are browsed and each column correspond to a dimension.

inter_bbo::Vector{Float64} :

List of black box outputs corresponding to intermediate states evaluated during the optimization process. Lines correspond to successive states that are browsed and each column correspond to an output (same order as defined in nomadParameters.output_types).

inter_bbo::Vector{Int64} :

List of black box evaluations numbers required to reach each of the states available in inter_states.

source